JavaScript Error
Where I got frustrated with javascript and had to take a break and move on for a bit. I dont know why it wont work. Then why the heck there is a blank spot in the first image i have no clue how the heck I did that so yea moving on for a bit.